+++ src/shared/commons/utils/date.utils.ts
@@ -70,4 +70,10 @@ export class DateUtils {
             return dateend;
         }
     }
+
+    public static formatDateTime(start: Date): string {
+        const formatOptionsLong = { year: 'numeric', month: '2-digit', day: '2-digit', hour: '2-digit', minute: '2-digit',
+            second: '2-digit' };
+        return start.toLocaleString('de-DE', formatOptionsLong);
+    }
 }